+++ src/Helper/Regex.php
@@ -4,5 +4,6 @@ namespace App\Helper;
 
 class Regex
 {
+    public const UUID_V4_CONTROLLER = '[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}';
     public const UUID_V4 = '/^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$/';
 }