+++ src/Type/AccessType.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Type;
+
+enum AccessType: string
+{
+    case READ = 'READ';
+}