+++ src/Type/ElementType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Type;
+
+enum ElementType: string
+{
+    case NODE = 'node';
+    case RELATION = 'relation';
+}